Move helper function into ARViewProvider
